Removed redefinition of VarMapMal and moved it to VarMapVal.h. (#1026)
This change makes the project perfectly compilable for C++20, everything works just fine.

+++ TombEngine/Specific/level.cpp
@@ -184,9 +184,9 @@ void LoadItems()
 			item->TriggerFlags = ReadInt16();
 			item->Flags = ReadInt16();
 			item->Name = ReadString();
-
-			g_GameScriptEntities->AddName(item->Name, i);
-			g_GameScriptEntities->TryAddColliding(i);
+			
+			g_GameScriptEntities->AddName(item->Name, static_cast<short>(i));
+			g_GameScriptEntities->TryAddColliding(static_cast<short>(i));
 
 			memcpy(&item->StartPose, &item->Pose, sizeof(Pose));
 		}
